Refactor packages under valorless.valorlessutils
Move color and cache classes into the valorless.valorlessutils.* packages (PlayerCache, SkinCache, Lang, McToAnsi) and update imports accordingly in ValorlessUtils and Log. Also bump buildNumber to 372. This cleans up package structure and ensures imports reference the new locations.
