+++ src/main/resources/buildNumber.properties
@@ -1,3 +1,3 @@
 #maven.buildNumber.plugin properties file
-#Thu Apr 23 03:41:59 CEST 2026
-buildNumber=371
+#Thu Apr 23 03:52:10 CEST 2026
+buildNumber=372
